test: split undo tag tests into dedicated module
- Move TestUndoRunTags and TestMessageIdsByLatestPriorRun out of
  test_tagging_integration.py into test_undo_run_tags.py, keeping both
  source modules under the CodeScene file-length threshold.
- Derive team from config in the _make_run/_make_run_with_results helpers
  to drop them from 5 to 4 arguments (CodeScene Excess Arguments).
